Refactor to reduce copy and paste code
