+++ slack/web/async_base_client.py
@@ -1,14 +1,16 @@
-import asyncio
-import json
 import logging
 from typing import Optional, Union
-from urllib.parse import urljoin
 
 import aiohttp
-from aiohttp import FormData, BasicAuth
+from aiohttp import FormData
 
-import slack.errors as err
-from slack.web import convert_bool_to_0_or_1, get_user_agent
+from slack.web import convert_bool_to_0_or_1
+from slack.web.async_internal_utils import (
+    _build_req_args,
+    _get_url,
+    _files_to_data,
+    _request_with_session,
+)
 from slack.web.async_slack_response import AsyncSlackResponse
 from slack.web.deprecation import show_2020_01_deprecation
 
@@ -21,7 +23,6 @@ class AsyncBaseClient:
         token=None,
         base_url=BASE_URL,
         timeout=30,
-        loop: Optional[asyncio.AbstractEventLoop] = None,
         ssl=None,
         proxy=None,
         trust_env_in_session: bool = False,
@@ -38,63 +39,6 @@ class AsyncBaseClient:
         self.trust_env_in_session = trust_env_in_session
         self.headers = headers or {}
         self._logger = logging.getLogger(__name__)
-        self._event_loop = loop
-
-    @staticmethod
-    def _get_event_loop():
-        """Retrieves the event loop or creates a new one."""
-        try:
-            return asyncio.get_event_loop()
-        except RuntimeError:
-            loop = asyncio.new_event_loop()
-            asyncio.set_event_loop(loop)
-            return loop
-
-    def _get_headers(
-        self,
-        *,
-        token: Optional[str],
-        has_json: bool,
-        has_files: bool,
-        request_specific_headers: Optional[dict],
-    ):
-        """Constructs the headers need for a request.
-        Args:
-            has_json (bool): Whether or not the request has json.
-            has_files (bool): Whether or not the request has files.
-            request_specific_headers (dict): Additional headers specified by the user for a specific request.
-
-        Returns:
-            The headers dictionary.
-                e.g. {
-                    'Content-Type': 'application/json;charset=utf-8',
-                    'Authorization': 'Bearer xoxb-1234-1243',
-                    'User-Agent': 'Python/3.6.8 slack/2.1.0 Darwin/17.7.0'
-                }
-        """
-        final_headers = {
-            "User-Agent": get_user_agent(),
-            "Content-Type": "application/x-www-form-urlencoded",
-        }
-
-        if token:
-            final_headers.update({"Authorization": "Bearer {}".format(token)})
-
-        # Merge headers specified at client initialization.
-        final_headers.update(self.headers)
-
-        # Merge headers specified for a specific request. e.g. oauth.access
-        if request_specific_headers:
-            final_headers.update(request_specific_headers)
-
-        if has_json:
-            final_headers.update({"Content-Type": "application/json;charset=utf-8"})
-
-        if has_files:
-            # These are set automatically by the aiohttp library.
-            final_headers.pop("Content-Type", None)
-
-        return final_headers
 
     async def api_call(  # skipcq: PYL-R1710
         self,
@@ -138,62 +82,26 @@ class AsyncBaseClient:
             SlackRequestError: Json data can only be submitted as
                 POST requests.
         """
-        has_json = json is not None
-        has_files = files is not None
-        if has_json and http_verb != "POST":
-            msg = "Json data can only be submitted as POST requests. GET requests should use the 'params' argument."
-            raise err.SlackRequestError(msg)
 
-        api_url = self._get_url(api_method)
-
-        if auth:
-            auth = BasicAuth(auth["client_id"], auth["client_secret"])
-
-        if data is not None and isinstance(data, dict):
-            data = {k: v for k, v in data.items() if v is not None}
-        if files is not None and isinstance(files, dict):
-            files = {k: v for k, v in files.items() if v is not None}
-        if params is not None and isinstance(params, dict):
-            params = {k: v for k, v in params.items() if v is not None}
-
-        token: Optional[str] = self.token
-        if params is not None and "token" in params:
-            token = params.pop("token")
-        if json is not None and "token" in json:
-            token = json.pop("token")
-        req_args = {
-            "headers": self._get_headers(
-                token=token,
-                has_json=has_json,
-                has_files=has_files,
-                request_specific_headers=headers,
-            ),
-            "data": data,
-            "files": files,
-            "params": params,
-            "json": json,
-            "ssl": self.ssl,
-            "proxy": self.proxy,
-            "auth": auth,
-        }
+        api_url = _get_url(self.base_url, api_method)
+        req_args = _build_req_args(
+            token=self.token,
+            http_verb=http_verb,
+            files=files,
+            data=data,
+            params=params,
+            json=json,  # skipcq: PYL-W0621
+            headers=headers,
+            auth=auth,
+            ssl=self.ssl,
+            proxy=self.proxy,
+        )
 
         show_2020_01_deprecation(api_method)
 
         return await self._send(
             http_verb=http_verb, api_url=api_url, req_args=req_args,
         )
-
-    def _get_url(self, api_method):
-        """Joins the base Slack URL and an API method to form an absolute URL.
-
-        Args:
-            api_method (str): The Slack Web API method. e.g. 'chat.postMessage'
-
-        Returns:
-            The absolute API URL.
-                e.g. 'https://www.slack.com/api/chat.postMessage'
-        """
-        return urljoin(self.base_url, api_method)
 
     # =================================================================
     # aiohttp based async WebClient
@@ -216,16 +124,7 @@ class AsyncBaseClient:
         Returns:
             The response parsed into a AsyncSlackResponse object.
         """
-        open_files = []
-        files = req_args.pop("files", None)
-        if files is not None:
-            for k, v in files.items():
-                if isinstance(v, str):
-                    f = open(v.encode("utf-8", "ignore"), "rb")
-                    open_files.append(f)
-                    req_args["data"].update({k: f})
-                else:
-                    req_args["data"].update({k: v})
+        open_files = _files_to_data(req_args)
 
         if "params" in req_args:
             # True/False -> "1"/"0"
@@ -251,37 +150,11 @@ class AsyncBaseClient:
         Returns:
             A dictionary of the response data.
         """
-        session = None
-        use_running_session = self.session and not self.session.closed
-        if use_running_session:
-            session = self.session
-        else:
-            session = aiohttp.ClientSession(
-                timeout=aiohttp.ClientTimeout(total=self.timeout),
-                auth=req_args.pop("auth", None),
-                trust_env=self.trust_env_in_session,
-            )
-
-        response = None
-        try:
-            async with session.request(http_verb, api_url, **req_args) as res:
-                data = {}
-                try:
-                    data = await res.json()
-                except aiohttp.ContentTypeError:
-                    self._logger.debug(
-                        f"No response data returned from the following API call: {api_url}."
-                    )
-                except json.decoder.JSONDecodeError as e:
-                    message = f"Failed to parse the response body: {str(e)}"
-                    raise err.SlackApiError(message, res)
-
-                response = {
-                    "data": data,
-                    "headers": res.headers,
-                    "status_code": res.status,
-                }
-        finally:
-            if not use_running_session:
-                await session.close()
-        return response
+        return await _request_with_session(
+            current_session=self.session,
+            timeout=self.timeout,
+            logger=self._logger,
+            http_verb=http_verb,
+            api_url=api_url,
+            req_args=req_args,
+        )
